Conflicts:
	app/controllers/editor/auth_renderer.rb
	app/helpers/cms_helper.rb
	app/models/mail_template.rb
	app/models/site_node.rb
	app/views/templates/edit.rhtml
	lib/tasks/install_module.rake
	spec/controllers/editor/publication_renderer_spec.rb
	spec/models/content_model_segment_field_spec.rb
	spec/models/content_publication_spec.rb
	spec/webiva_spec_helper.rb
	vendor/modules/blog/app/models/blog/blog_post.rb
	vendor/modules/blog/app/views/blog/manage/post.rhtml
	vendor/modules/user_profile/spec/controllers/user_profile/manage_controller_spec.rb
	vendor/plugins/validates_as_email/lib/validates_as_email.rb

+++ spec/webiva_spec_helper.rb
@@ -136,7 +136,8 @@ class DomainModel
     DomainModel.__reset_modified_tables
   end
 
-  before_create do
+  before_create :__add_table
+  def __add_table
     DomainModel.__add_modified_table self.class.table_name
   end
 end
@@ -171,7 +172,8 @@ class SystemModel
     SystemModel.__reset_modified_tables
   end
 
-  before_create do
+  before_create :__add_table
+  def __add_table
     SystemModel.__add_modified_table self.class.table_name
   end
 end
@@ -660,7 +662,7 @@ RSpec.configure do |config|
   end
 
   config.after(:suite) do
-    File.delete tests_are_running_file
+    File.delete tests_are_running_file if File.exists?(tests_are_running_file)
   end
 
   config.before(:each) do
@@ -678,4 +680,3 @@ RSpec.configure do |config|
     :file_path => /_feature_spec/
   }
 end
-
